Remove title attribute from a[href]
Prefer alt text in images to give an accessible name
where an icon is present. In other cases, where visible
text is available, the title does not contribute anything
and is a risk of the accessible name drifting from the
visible one. It also is unreliable as part of the accessible
name computation.